+++ src/server/game/Spells/SpellInfo.cpp
@@ -1490,7 +1490,7 @@ bool SpellInfo::IsAbilityOfSkillType(uint32 skillType) const
     SkillLineAbilityMapBounds bounds = sSpellMgr->GetSkillLineAbilityMapBounds(Id);
 
     for (SkillLineAbilityMap::const_iterator _spell_idx = bounds.first; _spell_idx != bounds.second; ++_spell_idx)
-        if (_spell_idx->second->SkillLine == skillType)
+        if (uint32(_spell_idx->second->SkillLine) == skillType)
             return true;
 
     return false;
