+++ src/server/game/Spells/SpellMgr.cpp
@@ -120,7 +120,7 @@ bool IsPartOfSkillLine(uint32 skillId, uint32 spellId)
 {
     SkillLineAbilityMapBounds skillBounds = sSpellMgr->GetSkillLineAbilityMapBounds(spellId);
     for (SkillLineAbilityMap::const_iterator itr = skillBounds.first; itr != skillBounds.second; ++itr)
-        if (itr->second->SkillLine == skillId)
+        if (uint32(itr->second->SkillLine) == skillId)
             return true;
 
     return false;
